Centralize page model
