+++ frontend/src/app/shared/models/page.model.ts
@@ -1,0 +1,7 @@
+export interface Page<T> {
+  content: T[];
+  totalElements: number;
+  totalPages: number;
+  size: number;
+  number: number;
+}
